final debugging

+++ controllers/screenShot.controller.js
@@ -1,15 +1,19 @@
 import path from 'path';
 import fs from 'fs';
-import { __dirname } from '../utils/pathHelper.js';
-import { captureScreenshot } from '../services/screenshotService.js';
+import { __dirname } from '../utils/pathHelper.util.js';
+import { captureScreenshot } from '../services/screenShot.service.js';
 
-// A function to handle screenshots.
 export async function handleScreenshot(req, res) {
-  const screenshotPath = path.join(__dirname, '..', 'tmp', 'screenshot.png');
-  const imagePageUrl = 'http://localhost:8000/preview';
+  const imageUrl = req.app.locals.imageUrl;
+
+  if (!imageUrl) {
+    return res.status(400).send("No image uploaded.");
+  }
+
+  const screenshotPath = path.join(__dirname, 'uploads', 'screenshot.png');
 
   try {
-    await captureScreenshot(imagePageUrl, screenshotPath);
+    await captureScreenshot(imageUrl, screenshotPath);
     res.download(screenshotPath, 'image-screenshot.png', () => {
       fs.unlink(screenshotPath, () => {});
     });
